# openapi/openapi.yml
openapi: 3.1.0
info:
  title: MLB Forecast API
  description: "\n    # MLB Predictions Backend API\n\n    Enterprise-grade MLB game\
    \ prediction platform delivering real-time analytics and machine learning insights.\n\
    \n    ## \U0001F680 Core Features\n\n    ### \u26BE **Data Ingestion Pipeline**\n\
    \    - Real-time game results from official MLB APIs\n    - Historical statistics\
    \ and performance metrics\n    - Live game schedules and league standings\n  \
    \  - Automated data validation and cleansing\n\n    ### \U0001F916 **Advanced\
    \ Machine Learning**\n    - Probabilistic game outcome predictions\n    - Total\
    \ runs and score forecasting\n    - Player performance analytics\n    - Continuous\
    \ model training and optimization\n\n    ### \u26A1 **High-Performance Architecture**\n\
    \    - Multi-layer Redis caching system\n    - Optimized database queries with\
    \ indexing\n    - Asynchronous processing for scalability\n    - Sub-second API\
    \ response times\n\n    ### \U0001F4CA **Analytics & Insights**\n    - Team performance\
    \ trending\n    - Head-to-head historical analysis\n    - Weather impact on game\
    \ outcomes\n    - Injury reports integration\n\n    ## \U0001F527 Technical Stack\n\
    \n    - **Framework**: FastAPI with async/await patterns\n    - **Database**:\
    \ PostgreSQL with optimized schemas\n    - **Cache**: Redis for high-speed data\
...
    \    ```\n    "
  version: 1.0.0
paths:
...
  /api/v1/data/ingest/team_stats:
    post:
      tags:
        - Teams
        - Stats
        - Data Ingestion
      summary: Ingest All Team Stats
      description: "Ingest all team statistics data from the MLB API for a specific\
        \ season.\n\nThis endpoint allows for the ingestion of all team statistics\
        \ (hitting, pitching, fielding, catching)\nfrom the MLB API for a given season.\
        \ It uses the IngestAllTeamStatsUseCase to fetch and store\nstatistics for\
        \ all teams in the specified season.\n\nParameters:\n    season (int): The\
        \ season year to ingest statistics for. Defaults to the current year.\n  \
        \  use_cases (Dict): Dependency-injected dictionary containing team stats\
        \ ingestion use cases.\n\nReturns:\n    dict: A dictionary containing success\
        \ status, message, and data about the ingested\n    team statistics including\
        \ the count and season.\n\nRaises:\n    HTTPException: If an error occurs\
        \ during the team statistics ingestion process,\n    an HTTP 500 status code\
        \ is raised with an error message."
      operationId: ingest_all_team_stats_api_v1_data_ingest_team_stats_post
      parameters:
        - name: season
          in: query
          required: false
          schema:
            type: integer
            description: Season to ingest team statistics for
            default: 2025
            title: Season
          description: Season to ingest team statistics for
      responses:
        '201':
          description: Ingest all team statistics data from the MLB API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamStatsIngestionResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
  /api/v1/data/ingest/team_stats/hitting:
    post:
      tags:
        - Teams
        - Stats
        - Data Ingestion
      summary: Ingest Team Hitting Stats
      description: "Ingest team hitting statistics data from the MLB API for a specific\
        \ season.\n\nThis endpoint allows for the ingestion of team hitting statistics\
        \ from the MLB API\nfor a given season. It uses the IngestTeamHittingStatsUseCase\
        \ to fetch and store\nhitting statistics for all teams in the specified season.\n\
        \nParameters:\n    season (int): The season year to ingest statistics for.\
        \ Defaults to the current year.\n    use_cases (Dict): Dependency-injected\
        \ dictionary containing team stats ingestion use cases.\n\nReturns:\n    dict:\
        \ A dictionary containing success status, message, and data about the ingested\n\
        \    team hitting statistics including the count and season.\n\nRaises:\n\
        \    HTTPException: If an error occurs during the team hitting statistics\
        \ ingestion process,\n    an HTTP 500 status code is raised with an error\
        \ message."
      operationId: ingest_team_hitting_stats_api_v1_data_ingest_team_stats_hitting_post
      parameters:
        - name: season
          in: query
          required: false
          schema:
            type: integer
            description: Season to ingest team hitting statistics for
            default: 2025
            title: Season
          description: Season to ingest team hitting statistics for
      responses:
        '201':
          description: Team hitting statistics ingested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamStatsIngestionResponse'
              example:
                status: success
                code: 201
                data:
                  hitting_stats_count: 30
                  season: 2025
                errors: []
                message: Team hitting statistics ingested successfully for season 2025
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
  /api/v1/data/ingest/team_stats/pitching:
    post:
      tags:
        - Teams
        - Stats
        - Data Ingestion
      summary: Ingest Team Pitching Stats
      description: "Ingest team pitching statistics data from the MLB API for a specific\
        \ season.\n\nThis endpoint allows for the ingestion of team pitching statistics\
        \ from the MLB API\nfor a given season. It uses the IngestTeamPitchingStatsUseCase\
        \ to fetch and store\npitching statistics for all teams in the specified season.\n\
        \nParameters:\n    season (int): The season year to ingest statistics for.\
        \ Defaults to the current year.\n    use_cases (Dict): Dependency-injected\
        \ dictionary containing team stats ingestion use cases.\n\nReturns:\n    dict:\
        \ A dictionary containing success status, message, and data about the ingested\n\
        \    team pitching statistics including the count and season.\n\nRaises:\n\
        \    HTTPException: If an error occurs during the team pitching statistics\
        \ ingestion process,\n    an HTTP 500 status code is raised with an error\
        \ message."
      operationId: ingest_team_pitching_stats_api_v1_data_ingest_team_stats_pitching_post
      parameters:
        - name: season
          in: query
          required: false
          schema:
            type: integer
            description: Season to ingest team pitching statistics for
            default: 2025
            title: Season
          description: Season to ingest team pitching statistics for
      responses:
        '201':
          description: Team pitching statistics ingested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamStatsIngestionResponse'
              example:
                status: success
                code: 201
                data:
                  pitching_stats_count: 30
                  season: 2025
                errors: []
                message: Team pitching statistics ingested successfully for season 2025
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
  /api/v1/data/ingest/team_stats/fielding:
    post:
      tags:
        - Teams
        - Stats
        - Data Ingestion
      summary: Ingest Team Fielding Stats
      description: "Ingest team fielding statistics data from the MLB API for a specific\
        \ season.\n\nThis endpoint allows for the ingestion of team fielding statistics\
        \ from the MLB API\nfor a given season. It uses the IngestTeamFieldingStatsUseCase\
        \ to fetch and store\nfielding statistics for all teams in the specified season.\n\
        \nParameters:\n    season (int): The season year to ingest statistics for.\
        \ Defaults to the current year.\n    use_cases (Dict): Dependency-injected\
        \ dictionary containing team stats ingestion use cases.\n\nReturns:\n    dict:\
        \ A dictionary containing success status, message, and data about the ingested\n\
        \    team fielding statistics including the count and season.\n\nRaises:\n\
        \    HTTPException: If an error occurs during the team fielding statistics\
        \ ingestion process,\n    an HTTP 500 status code is raised with an error\
        \ message."
      operationId: ingest_team_fielding_stats_api_v1_data_ingest_team_stats_fielding_post
      parameters:
        - name: season
          in: query
          required: false
          schema:
            type: integer
            description: Season to ingest team fielding statistics for
            default: 2025
            title: Season
          description: Season to ingest team fielding statistics for
      responses:
        '201':
          description: Team fielding statistics ingested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamStatsIngestionResponse'
              example:
                status: success
                code: 201
                data:
                  fielding_stats_count: 30
                  season: 2025
                errors: []
                message: Team fielding statistics ingested successfully for season 2025
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
  /api/v1/data/ingest/team_stats/catching:
    post:
      tags:
        - Teams
        - Stats
        - Data Ingestion
        - Teams
        - Stats
        - Data Ingestion
      summary: Ingest Team Catching Stats
      description: "Ingest team catching statistics data from the MLB API for a specific\
        \ season.\n\nThis endpoint allows for the ingestion of team catching statistics\
        \ from the MLB API\nfor a given season. It uses the IngestTeamCatchingStatsUseCase\
        \ to fetch and store\ncatching statistics for all teams in the specified season.\n\
        \nParameters:\n    season (int): The season year to ingest statistics for.\
        \ Defaults to the current year.\n    use_cases (Dict): Dependency-injected\
        \ dictionary containing team stats ingestion use cases.\n\nReturns:\n    dict:\
        \ A dictionary containing success status, message, and data about the ingested\n\
        \    team catching statistics including the count and season.\n\nRaises:\n\
        \    HTTPException: If an error occurs during the team catching statistics\
        \ ingestion process,\n    an HTTP 500 status code is raised with an error\
        \ message."
      operationId: ingest_team_catching_stats_api_v1_data_ingest_team_stats_catching_post
      parameters:
        - name: season
          in: query
          required: false
          schema:
            type: integer
            description: Season to ingest team catching statistics for
            default: 2025
            title: Season
          description: Season to ingest team catching statistics for
      responses:
        '201':
          description: Team catching statistics ingested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamStatsIngestionResponse'
              example:
                status: success
                code: 201
                data:
                  catching_stats_count: 30
                  season: 2025
                errors: []
                message: Team catching statistics ingested successfully for season 2025
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
...
components:
  schemas:
...
    BadRequest:
      description: Bad Request - Invalid input parameters
      title: Bad Request
      $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found - Resource not found
      title: Not Found
      $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Validation Error - Request validation failed
      title: Validation Error
      $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal Server Error
      title: Internal Server Error
      $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Service Unavailable - External service error
      title: Service Unavailable
      $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      properties:
        status:
          type: string
          title: Status
          description: Response status
          default: error
        code:
          type: integer
          title: Code
          description: HTTP status code
        data:
          type: 'null'
          title: Data
          description: No data in error responses
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: List of error messages
        message:
          anyOf:
            - type: string
            - type: 'null'
          title: Message
          description: Error message
      type: object
      required:
        - code
        - errors
      title: ErrorResponse
      description: Error response model for API documentation.
      examples:
        - code: 400
          errors:
            - League must be 'American' or 'National'
          message: Invalid data provided
          status: error
        - code: 404
          errors:
            - Team with ID 999 not found
          message: Resource not found
          status: error
        - code: 500
          errors:
            - An unexpected error occurred
          message: Internal server error
          status: error
    TeamStatsIngestionData:
      properties:
        hitting_stats_count:
          type: integer
          title: Hitting stats created or updated
          description: Quantity of hitting stats created or updated
        pitching_stats_count:
          type: integer
          title: Pitching stats created or updated
          description: Quantity of pitching stats created or updated
        fielding_stats_count:
          type: integer
          title: Fielding stats created or updated
          description: Quantity of fielding stats created or updated
        catching_stats_count:
          type: integer
          title: Catching stats created or updated
          description: Quantity of catching stats created or updated
        season:
          type: integer
          title: Year of season
          description: Year of season from teams stats ingested
...
    TeamStatsIngestionResponse:
      properties:
        status:
          type: string
          title: Status
          default: success
        code:
          type: integer
          title: Code
          description: HTTP status code
        data:
          $ref: '#/components/schemas/TeamStatsIngestionData'
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: List of error messages
        message:
          type: string
          title: Message
          default: Full data ingestion completed successfully
      type: object
      required:
        - code
        - data
      title: TeamStatsIngestionResponse
      description: Response model for teams stats ingestion endpoint.
...